Added file download time calculation method.

+++ src/base/BaseTest.java
@@ -123,6 +123,9 @@ public  class BaseTest
 	public void fileDelete()
 	{
 		File f = new File(Constants.FILE_PATH);
-		f.delete();
+		if(fileExist())
+		{
+			f.delete();
+		}
 	}
 }
